feat: Display error for file replace
