+++ src/app/search/AudioResetForm.jsx
@@ -1,6 +1,7 @@
 import T from 'prop-types';
 
 import {
+  Box,
   Button,
   Modal,
   ModalOverlay,
@@ -14,18 +15,19 @@ import {
 import { MdChevronLeft } from 'react-icons/md';
 
 import { ACCEPTED_AUDIO_TYPES } from '@/settings';
+import { Error } from '@/components';
+import useUploadValidation from './hooks/useUploadValidation';
 
 export default function AudioResetForm({ setFile }) {
   const { isOpen, onOpen, onClose } = useDisclosure();
+  const { error, handleFileChange } = useUploadValidation(setFile);
 
   const handleReset = () => {
     onClose();
     const inputEl = document.createElement('input');
     inputEl.setAttribute('type', 'file');
     inputEl.setAttribute('accept', ACCEPTED_AUDIO_TYPES.join(','));
-    inputEl.addEventListener('change', (e) => {
-      setFile(e.target.files[0]);
-    });
+    inputEl.addEventListener('change', handleFileChange);
     inputEl.click();
   };
 
@@ -49,6 +51,12 @@ export default function AudioResetForm({ setFile }) {
         </ModalContent>
       </Modal>
       <Button type="button" variant="link" size="xs" leftIcon={<MdChevronLeft />} onClick={onOpen}>Upload new file</Button>
+      {error && (
+        <>
+          <Error>{ error }</Error>
+          <Box borderBottom="1px solid" borderBottomColor="neutral.100" my="4" />
+        </>
+      )}
     </>
   );
 }
