+++ src/app/search/AudioSelectForm.jsx
@@ -2,38 +2,18 @@ import { useRef, useState } from 'react';
 import T from 'prop-types';
 import { Box, Button, Flex, Text } from '@chakra-ui/react';
 
-import { ACCEPTED_AUDIO_TYPES, MAX_AUDIO_LENGTH, MAX_AUDIO_SIZE } from '@/settings';
+import { ACCEPTED_AUDIO_TYPES } from '@/settings';
 import { Error } from '@/components';
-import getDuration from '@/utils/getDuration';
+import useUploadValidation from './hooks/useUploadValidation';
 
 export default function AudioSelectForm({ handleFileSelect }) {
   const inputRef = useRef();
-  const [ error, setError ] = useState();
   const [ name, setName ] = useState();
-
-  const validate = async (file) => {
-    const duration = await getDuration(file);
-    const errors = [];
-    if (duration > MAX_AUDIO_LENGTH) {
-      errors.push('The audio length exceeds the limit of 5 minutes. Upload a shorter recording.');
-    }
-    if (file.size > MAX_AUDIO_SIZE) {
-      errors.push('The file size exceeds the limit of 1GB. Upload a smaller file.');
-    }
-    return errors;
-  };
+  const { error, handleFileChange: validate } = useUploadValidation(handleFileSelect);
 
   const handleFileChange = async (e) => {
-    const file = e.target.files[0];
-    setName(file.name);
-
-    const errors = await validate(file);
-    if (errors.length > 0) {
-      setError(errors.join(' '));
-    } else {
-      handleFileSelect(file);
-    }
-
+    setName(e.target.files[0].name);
+    validate(e);
   };
 
   return (
